Explicit conversion for type comparison issue

+++ src/melee/ft/ft_0881.c
@@ -150,7 +150,7 @@ static inline void inlineB0(Fighter* fp, s32 move_id)
 
 void ft_800890D0(Fighter* fp, u32 move_id)
 {
-    if (move_id == 1 || move_id != fp->x2068_attackID) {
+    if (move_id == 1 || move_id != (u32) fp->x2068_attackID) {
         inlineB0(fp, move_id);
     }
 }
@@ -161,7 +161,7 @@ void ft_800890D0(Fighter* fp, u32 move_id)
 
 static void inlineC0(Fighter* fp, u32 move_id)
 {
-    if (move_id == fp->x2068_attackID) {
+    if (move_id == (u32) fp->x2068_attackID) {
         inlineB0(fp, move_id);
     }
 }
